Additions to Card Class
Created the checkForBingo() method as well as the cardFreeze() and
cardWin() method. checkForBingo is almost done it just need to be able
to compare to the possible bingos the computer actually has. I need a
sit down with Greg to go over the freeze and win methods to have it
align with a label that will appear when either of them is called.