Added Credits Window
This is just a png image as one JLabel background. I figured it was
easier than adding in rendered on the fly text. It is modal so nothing
else can be done while using it. Also added resource files for Credits
as well as the psd file to edit the credits window if needed.