Added Shop Interface
This shop interface is almost completely functional. I have changed how
the game flow works slightly in that Game is gone and Bingo has the
main. Bingo’s main calls a new shop window which, when Start is
clicked, calls a new Bingo game instead of Bingo main calling a new
Bingo game right off the bat. The shop’s buttons are all done in code
but the rest is just rendered with Photoshop and the psd is in the
Source folder to edit if need be. Currently I’m having trouble with the
font not cutting off on number of cards over 2. I also could not get
getGameFont to work correctly so it is using a separate Cooper Black
font. If Greg wants to try to figure it out that would be great.