Updated Splash
The splash only displays on slow computers so mostly no one will ever
see it. But, for those who have slow computers, at least they have
something to look at! Also, it does not require the JVM to be active so
it does not slow down the launching of the app whatsoever.