Player Class Changes
I removed the Player class and made both Human and Computer not extend
anything. I moved the total players integer to the Bingo class and made
Human accept a number of cards to be purchased. I greatly simplified
the Human class as they do not need to have cards. All a human has is
tickets. The Shop class now also correctly adjusts a user’s ticket bank
when pressing Start.