Check for bing
The check for bingo method is working now. It first checks if a token
is placed, if it is it will take the value in the cell and checks if it
has been called

****** PLEASE DO NOT CHANGE THIS CODE WITHOUT SPEAKING WITH ME FIRST, I
WILL NEED TO COME BACK TO IT AND I WANT TO EASILY RECOGNIZE WHAT I DID
LAST, please and thank you.

If everyone could run it a few times and make sure it is working
appropriately id appreciate it. I tested it a lot but I might have
missed something.