Game Plays Itself!
The game finally plays itself with the computer players accurately
marking and checking for Bingos. The ratio of Bingos available could be
tweaked and we have yet to implement response time which could be
critical towards the end of the round since it goes very fast. Also
implemented a closeBingo method to tidy up some resources for garbage
collection when ending a round.