Began Round Summary Screen
Right now it shows up after the last Bingo is called and says no
tickets were won. It has a button that closes the window and takes you
to the shop. We need a shop button, for the correct amount of tickets
to be displayed, and eventually a ticket bank history graph.