Bank History & Round Summary
I changed the way Human works so that there is now a proper history
being kept of the ticket balance. For now, this is only to make the
Round Summary window display the proper amount of tickets won but it
can also be used to create a ticket bank histograph.