Implemented Player Persistence
When a new player is made, it checks to see if a properties file
already exists in the user’s home folder on their computer. If it does,
it uses that value, if not it uses 20. Whenever the game exits
completely, a hook catches it and writes to the properties file in the
user’s home folder.