currentBingoNumber Correctly Updated
Fixed it so that currentBingoNumber in NewBingoNotification is updated
when the player wins a Bingo on one of his/her cards. It does not show
a notification and I verified through testing. I accomplished that by
setting currentBingoNumber to public static and incrementing it
directly through PlayerCard. Also tweaked the Rules since there were
some errors.