MessageDialog Class
I created a generic MessageDialog class that has one label for a
message and one button. These are both part of the default constructor
and must be passed. Right now the text sizing is pretty poorly
implemented but it works well enough to display messages. It would be
good to create a more robust way of resizing text based on message
length as well as implementing a way to make sure the text is not only
centered, but middle aligned between the button and top of the dialog
window.