Implemented Better Text Resizing
The MessageDialog class now dynamically changes the font size used
depending on how long the message is. It works OK up to about 70 or so
characters and more than that it starts to make the font smaller than
anticipated. I’m not entirely sure why. It works well enough though and
at least all of the message will be displayed regardless of its length.
Also set Shop to be made after background music starts in case there is
a modal dialog that pops up on first run.