Delete Player / Close Button Implemented
You can now delete a player from the list. There is also a close button
but its just the Bingo token in the top left corner of the ChoosePlayer
and NewPlayer dialogs. Moved writing of and deleting of players to the
shutdown hook only. Still need to make sure Command + Q on a Mac
invokes the shutdown hook.