Squashed Two ChoosePlayer/NewPlayer Bugs
1. When starting with no players, clicking the close button would take
you to the shop with no player name set. Now it just exits the program
as expected.
2. If you added a new player from the ChoosePlayer dialog and then
closed the program with the close button, the app would hang. Now it
exits and saves the new player(s) you made, even if they weren’t used.

I also changed the Shop interface slightly to move the ticket bank to
the right a little and moved the SHOP title down a bit to make room for
a message letting you know which player you’re using.