Implemented Reset Ticket Bank Button
It shows up if you get to four or fewer tickets after a round has ended
or if you start off a game with only that few tickets. It disappears as
soon as you click it. Right now it’s a generic button but Sidney will
design a button to take its place.