Fixed ChoosePlayer Bug/Removed Message
There was a bug where if you had deleted all the players in the list
and pressed start it would show a NewPlayer dialog and if you clicked
the close button it started the game with null as the name. Now, if you
click start nothing happens, suggesting that you should add a player.
May change it so that a new player dialog actually works.