NewPlayer In Center If First Window
It was offset before. Now NewPlayer takes a boolean and any place that
calls it can say if it is the first window or not. If it is, just set
location relative to null and if not, then move it over for Windows 7
computers to not experience the missing background image issue.